fix: adjust after comments
