+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "elephant-web",
+  "name": "elephant-chrome",
   "private": false,
   "license": "MIT",
   "version": "0.0.0",
@@ -14,7 +14,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@ttab/elephant-ui": "^0.0.4",
+    "@ttab/elephant-ui": "^0.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "uuid": "^9.0.1"
