+++ src/components/Link/index.tsx
@@ -1,3 +1,4 @@
+import type { MouseEvent } from 'react'
 import { NavigationActionType, type View } from '@/types'
 import { useNavigation } from '@/hooks'
 import { v4 as uuid } from 'uuid'
@@ -12,32 +13,40 @@ export const Link = ({ children, to, props }: LinkProps): JSX.Element => {
   const { state, dispatch } = useNavigation()
   const id = uuid()
   const linkItem = state.registry.get(to)
+
+  const handleClick = (event: MouseEvent<HTMLAnchorElement>): void => {
+    if (event.ctrlKey || event.metaKey) {
+      return
+    }
+
+    event.preventDefault()
+
+    dispatch({
+      type: NavigationActionType.ADD,
+      component: linkItem.component,
+      props: { ...props, id }
+    })
+
+    history.pushState({
+      id,
+      props: { ...props, id },
+      itemName: linkItem.metadata.name,
+      contentState: [
+        ...history.state.contentState,
+        {
+          id,
+          name: linkItem.metadata.name,
+          props
+        }
+      ]
+    }, children, `${linkItem.metadata.path}?id=${id}`)
+  }
+
   return (
     <a
       className='p-1 hover:font-bold'
       href={`${linkItem.metadata.path}?id=${id}`}
-      onClick={(event) => {
-        event.preventDefault()
-        dispatch({
-          type: NavigationActionType.ADD,
-          component: linkItem.component,
-          props: { ...props, id }
-        })
-
-        history.pushState({
-          id,
-          props: { ...props, id },
-          itemName: linkItem.metadata.name,
-          contentState: [
-            ...history.state.contentState,
-            {
-              id,
-              name: linkItem.metadata.name,
-              props
-            }
-          ]
-        }, children, `${linkItem.metadata.path}?id=${id}`)
-      }}
+      onClick={handleClick}
     >
       {children}
     </a>
